requirements.txt
imagehash
pillow
tqdm
configparser
flask